requirements.txt
flask==2.3.2
scrapy==2.9.0
psycopg2-binary==2.9.6
